+++ src/ansys/speos/core/project.py
@@ -23,17 +23,17 @@
 
 from __future__ import annotations
 
-import os
+from pathlib import Path
 import re
-from typing import List, Mapping, Optional, Union
+from typing import TYPE_CHECKING, List, Mapping, Optional, Union
 import uuid
 
 from google.protobuf.internal.containers import RepeatedScalarFieldContainer
 import numpy as np
-import pyvista as pv
 
 import ansys.speos.core.body as body
 import ansys.speos.core.face as face
+from ansys.speos.core.generic.general_methods import graphics_required
 from ansys.speos.core.kernel.body import BodyLink
 from ansys.speos.core.kernel.face import FaceLink
 from ansys.speos.core.kernel.part import ProtoPart
@@ -57,6 +57,18 @@ from ansys.speos.core.source import (
     SourceSurface,
 )
 from ansys.speos.core.speos import Speos
+
+try:
+    from ansys.speos.core.generic.general_methods import run_if_graphics_required
+
+    run_if_graphics_required(warning=True)
+except ImportError as err:  # pragma: no cover
+    raise err
+
+if TYPE_CHECKING:  # pragma: no cover
+    import pyvista as pv
+
+    from ansys.tools.visualization_interface import Plotter
 
 
 class Project:
@@ -735,6 +747,7 @@ class Project:
                 op_feature._fill(mat_inst=mat_inst)
 
         for src_inst in scene_data.sources:
+            src_feat = None
             if src_inst.HasField("rayfile_properties"):
                 src_feat = SourceRayFile(
                     project=self,
@@ -756,9 +769,11 @@ class Project:
                     source_instance=src_inst,
                     default_values=False,
                 )
-            self._features.append(src_feat)
+            if src_feat is not None:
+                self._features.append(src_feat)
 
         for ssr_inst in scene_data.sensors:
+            ssr_feat = None
             if ssr_inst.HasField("irradiance_properties"):
                 ssr_feat = SensorIrradiance(
                     project=self,
@@ -783,6 +798,7 @@ class Project:
             self._features.append(ssr_feat)
 
         for sim_inst in scene_data.simulations:
+            sim_feat = None
             simulation_template_link = self.client[sim_inst.simulation_guid].get()
             if simulation_template_link.HasField("direct_mc_simulation_template"):
                 sim_feat = SimulationDirect(
@@ -828,6 +844,7 @@ class Project:
         pv.PolyData
             mesh data extracted.
         """
+        import pyvista as pv
 
         def local2absolute(local_vertice: np.ndarray, coordinates) -> np.ndarray:
             """Convert local coordinate to global coordinate.
@@ -887,7 +904,45 @@ class Project:
                     part_mesh_info = part_mesh_info.append_polydata(face_mesh_data)
         return part_mesh_info
 
-    def _create_preview(self, viz_args=None) -> pv.Plotter:
+    def _create_speos_feature_preview(
+        self, plotter: Plotter, speos_feature: Union[SensorCamera, SensorRadiance, SensorIrradiance]
+    ) -> Plotter:
+        """Add speos feature visual preview to pyvista plotter object.
+
+        Parameters
+        ----------
+        plotter: Plotter
+            ansys.tools.visualization_interface.Plotter
+        speos_feature: Union[SensorCamera, SensorRadiance, SensorIrradiance]
+            speos feature whose visual data will be added.
+
+        Returns
+        -------
+        Plotter
+            ansys.tools.visualization_interface.Plotter
+        """
+        if not isinstance(speos_feature, (SensorIrradiance, SensorRadiance, SensorCamera)):
+            return plotter
+        plotter.plot(
+            speos_feature.visual_data.data,
+            show_edges=True,
+            line_width=2,
+            edge_color="red",
+            color="orange",
+            opacity=0.5,
+        )
+        match speos_feature:
+            case SensorRadiance():
+                plotter.plot(speos_feature.visual_data.coordinates.x_axis, color="red")
+                plotter.plot(speos_feature.visual_data.coordinates.y_axis, color="green")
+            case SensorIrradiance() | SensorCamera():
+                plotter.plot(speos_feature.visual_data.coordinates.x_axis, color="red")
+                plotter.plot(speos_feature.visual_data.coordinates.y_axis, color="green")
+                plotter.plot(speos_feature.visual_data.coordinates.z_axis, color="blue")
+        return plotter
+
+    @graphics_required
+    def _create_preview(self, viz_args=None) -> Plotter:
         """Create preview pyvista plotter object.
 
         Parameters
@@ -899,8 +954,13 @@ class Project:
             - {'style': 'surface', 'color':'white'},
             - {'opacity': 0.7, 'color':'white', 'show_edges': False},
         """
+        import pyvista as pv
+
+        from ansys.tools.visualization_interface import Plotter
+
         if viz_args is None:
             viz_args = {}
+
         _preview_mesh = pv.PolyData()
         # Retrieve root part
         root_part_data = self.client[self.scene_link.get().part_guid].get()
@@ -920,11 +980,21 @@ class Project:
         poly_data = self.__extract_part_mesh_info(part_data=root_part_data)
         if poly_data is not None:
             _preview_mesh = _preview_mesh.append_polydata(poly_data)
-        p = pv.Plotter()
-        p.add_mesh(_preview_mesh, show_edges=True, **viz_args)
+        p = Plotter()
+        viz_args["show_edges"] = True
+        p.plot(_preview_mesh, **viz_args)
+
+        # Add speos visual data at the root part
+        for feature in self._features:
+            p = self._create_speos_feature_preview(p, feature)
         return p
 
-    def preview(self, viz_args=None) -> None:
+    @graphics_required
+    def preview(
+        self,
+        viz_args=None,
+        screenshot: Optional[Union[str, Path]] = None,
+    ) -> None:
         """Preview cad bodies inside the project's scene.
 
         Parameters
@@ -934,12 +1004,17 @@ class Project:
             e.g.
             - {'style': 'wireframe'},
             - {'style': 'surface', 'color':'white'},
-            - {'opacity': 0.7, 'color':'white', 'show_edges': False},
+            - {'opacity': 0.7, 'color':'white', 'show_edges': False}.
+
+        screenshot : str or Path or ``None``
+            Path to save a screenshot of the plotter. If defined Plotter will only create the
+            screenshot
+
         """
         if viz_args is None:
             viz_args = {"opacity": 1}
+        if screenshot is not None:
+            screenshot = Path(screenshot)
+
         p = self._create_preview(viz_args=viz_args)
-        if os.environ.get("DOCUMENTATION_BUILDING", "true") == "true":
-            p.show(jupyter_backend="html")
-        else:
-            p.show()
+        p.show(screenshot=screenshot)