+++ tests/core/test_opt_prop.py
@@ -146,11 +146,21 @@ def test_reset_optical_property(speos: Speos):
     """Test reset of optical property."""
     p = Project(speos=speos)
 
+    root_part = p.create_root_part()
+    body_b = root_part.create_body(name="TheBodyB")
+    (
+        body_b.create_face(name="TheFaceF")
+        .set_vertices([0, 0, 0, 1, 0, 0, 0, 1, 0])
+        .set_facets([0, 1, 2])
+        .set_normals([0, 0, 1, 0, 0, 1, 0, 0, 1])
+    )
+    root_part.commit()
+
     # Create + commit
     op1 = (
         p.create_optical_property(name="Material.1")
         .set_volume_opaque()
-        .set_geometries(geometries=[GeoRef.from_native_link("TheBodyB")])
+        .set_geometries(geometries=[body_b.geo_path])
         .commit()
     )
     assert op1.vop_template_link is not None
@@ -185,6 +195,13 @@ def test_delete_optical_property(speos: Speos):
     """Test delete of optical property."""
     p = Project(speos=speos)
 
+    root_part = p.create_root_part()
+    body_b = root_part.create_body(name="TheBodyB")
+    body_b.create_face(name="TheFaceF").set_vertices([0, 0, 0, 1, 0, 0, 0, 1, 0]).set_facets(
+        [0, 1, 2]
+    ).set_normals([0, 0, 1, 0, 0, 1, 0, 0, 1])
+    root_part.commit()
+
     # Create + commit
     op1 = (
         p.create_optical_property(name="Material.1")
@@ -216,10 +233,25 @@ def test_delete_optical_property(speos: Speos):
 def test_get_optical_property(speos: Speos, capsys):
     """Test get of an optical property."""
     p = Project(speos=speos)
+
+    root_part = p.create_root_part()
+    body_a = root_part.create_body(name="TheBodyA")
+    face = (
+        body_a.create_face(name="TheFaceF")
+        .set_vertices([0, 0, 0, 1, 0, 0, 0, 1, 0])
+        .set_facets([0, 1, 2])
+        .set_normals([0, 0, 1, 0, 0, 1, 0, 0, 1])
+    )
+    body_b = root_part.create_body(name="TheBodyB")
+    body_b.create_face(name="TheFaceF").set_vertices([0, 0, 0, 1, 0, 0, 0, 1, 0]).set_facets(
+        [0, 1, 2]
+    ).set_normals([0, 0, 1, 0, 0, 1, 0, 0, 1])
+    root_part.commit()
+
     op1 = (
         p.create_optical_property(name="Material.1")
         .set_volume_opaque()
-        .set_geometries(geometries=[GeoRef.from_native_link("TheBodyA")])
+        .set_geometries(geometries=[body_a])
         .commit()
     )
 
@@ -243,7 +275,7 @@ def test_get_optical_property(speos: Speos, capsys):
         p.create_optical_property(name="OpticalProperty2")
         .set_volume_optic(index=1.7, absorption=0.01, constringence=55)
         .set_surface_opticalpolished()
-        .set_geometries(geometries=[GeoRef.from_native_link("TheBodyB")])
+        .set_geometries(geometries=[body_b])
         .commit()
     )
 
@@ -266,6 +298,7 @@ def test_get_optical_property(speos: Speos, capsys):
         p.create_optical_property(name="OpticalProperty3")
         .set_volume_none()
         .set_surface_library(path=str(Path(test_path) / "R_test.anisotropicbsdf"))
+        .set_geometries(geometries=[face])
         .commit()
     )