+++ tests/core/test_part.py
@@ -51,7 +51,7 @@ def test_create_body(speos: Speos):
     # Add empty body
     body1 = root_part.create_body(name="Body.1").commit()
     assert len(root_part._geom_features) == 1
-
+    assert body1.geo_path.metadata["GeoPath"] == "Body.1"
     assert len(root_part.part_link.get().body_guids) == 1
     assert root_part.part_link.get().body_guids[0] == body1.body_link.key
 
@@ -59,7 +59,7 @@ def test_create_body(speos: Speos):
     body2 = root_part.create_body(name="Body.2")
     root_part.commit()
     assert len(root_part._geom_features) == 2
-
+    assert body2.geo_path.metadata["GeoPath"] == "Body.2"
     assert len(root_part.part_link.get().body_guids) == 2
     assert root_part.part_link.get().body_guids[1] == body2.body_link.key
 
@@ -75,10 +75,16 @@ def test_create_face(speos: Speos):
     p = Project(speos=speos)
     root_part = p.create_root_part()
     body1 = root_part.create_body(name="Body.1")
+    face0 = (
+        body1.create_face(name="TheFaceF")
+        .set_vertices([0, 0, 0, 1, 0, 0, 0, 1, 0])
+        .set_facets([0, 1, 2])
+        .set_normals([0, 0, 1, 0, 0, 1, 0, 0, 1])
+    )
     root_part.commit()
-    assert len(body1._geom_features) == 0
-    assert len(body1.body_link.get().face_guids) == 0
-
+    assert len(body1._geom_features) == 1
+    assert len(body1.body_link.get().face_guids) == 1
+    assert face0.geo_path.metadata["GeoPath"] == "Body.1/TheFaceF"
     # Add a face
     face1 = (
         body1.create_face(name="Face.1")
@@ -87,10 +93,10 @@ def test_create_face(speos: Speos):
         .set_normals([0, 0, 1, 0, 0, 1, 0, 0, 1])
         .commit()
     )
-    assert len(body1._geom_features) == 1
-
-    assert len(body1.body_link.get().face_guids) == 1
-    assert body1.body_link.get().face_guids[0] == face1.face_link.key
+    assert len(body1._geom_features) == 2
+    assert face1.geo_path.metadata["GeoPath"] == "Body.1/Face.1"
+    assert len(body1.body_link.get().face_guids) == 2
+    assert body1.body_link.get().face_guids[1] == face1.face_link.key
     assert face1.face_link.get().vertices == [0, 1, 0, 0, 2, 0, 1, 2, 0]
     assert face1.face_link.get().facets == [0, 1, 2]
     assert face1.face_link.get().normals == [0, 0, 1, 0, 0, 1, 0, 0, 1]
@@ -103,15 +109,16 @@ def test_create_face(speos: Speos):
         .set_normals([0, 0, 1, 0, 0, 1, 0, 0, 1])
         .commit()
     )
-    assert len(body1._geom_features) == 2
-
-    assert len(body1.body_link.get().face_guids) == 2
-    assert body1.body_link.get().face_guids[1] == face2.face_link.key
+    assert len(body1._geom_features) == 3
+    assert face2.geo_path.metadata["GeoPath"] == "Body.1/Face.2"
+    assert len(body1.body_link.get().face_guids) == 3
+    assert body1.body_link.get().face_guids[2] == face2.face_link.key
     assert face2.face_link.get().vertices == [0, 0, 0, 1, 0, 0, 0, 1, 0]
     assert face2.face_link.get().facets == [0, 2, 1]
     assert face2.face_link.get().normals == [0, 0, 1, 0, 0, 1, 0, 0, 1]
 
     # Delete faces
+    face0.delete()
     face1.delete()
     face2.delete()
     assert len(body1.body_link.get().face_guids) == 0
@@ -133,6 +140,7 @@ def test_create_subpart(speos: Speos):
     )
     assert len(root_part._geom_features) == 1
     assert len(root_part.part_link.get().parts) == 1
+    assert sp1.geo_path.metadata["GeoPath"] == "SubPart.1"
     assert root_part.part_link.get().parts[0] == sp1._part_instance
     assert sp1._part_instance.axis_system == [
         5,
@@ -161,6 +169,7 @@ def test_create_subpart(speos: Speos):
     root_part.commit()
     assert len(root_part._geom_features) == 2
     assert len(root_part.part_link.get().parts) == 2
+    assert sp2.geo_path.metadata["GeoPath"] == "SubPart.2"
     assert root_part.part_link.get().parts[1] == sp2._part_instance
     assert sp2._part_instance.axis_system == [
         15,
@@ -204,7 +213,7 @@ def test_create_subpart_body(speos: Speos):
     # Add empty body
     body1 = sp1.create_body(name="Body.1").commit()
     assert len(sp1._geom_features) == 1
-
+    assert body1.geo_path.metadata["GeoPath"] == "SubPart.1/Body.1"
     assert len(sp1.part_link.get().body_guids) == 1
     assert sp1.part_link.get().body_guids[0] == body1.body_link.key
 
@@ -212,7 +221,7 @@ def test_create_subpart_body(speos: Speos):
     body2 = sp1.create_body(name="Body.2")
     root_part.commit()
     assert len(sp1._geom_features) == 2
-
+    assert body2.geo_path.metadata["GeoPath"] == "SubPart.1/Body.2"
     assert len(sp1.part_link.get().body_guids) == 2
     assert sp1.part_link.get().body_guids[1] == body2.body_link.key
 
@@ -259,6 +268,7 @@ def test_create_subpart_subpart(speos: Speos):
         1,
     ]
     assert sp11._part_instance.part_guid == sp11.part_link.key
+    assert sp11.geo_path.metadata["GeoPath"] == "SubPart.1/SubPart.11"
     assert (
         sp11.part_link.get().name == "SubPart.11"
     )  # part contained in a sub part gets same name as the sub part
@@ -287,6 +297,7 @@ def test_create_subpart_subpart(speos: Speos):
         1,
     ]
     assert sp12._part_instance.part_guid == sp12.part_link.key
+    assert sp12.geo_path.metadata["GeoPath"] == "SubPart.1/SubPart.12"
     assert (
         sp12.part_link.get().name == "SubPart.12"
     )  # part contained in a sub part gets same name as the sub part
@@ -310,7 +321,10 @@ def test_commit_part(speos: Speos):
     sp1.create_sub_part(name="SubPart.11").set_axis_system(
         axis_system=[-5, -4, -10, 1, 0, 0, 0, 1, 0, 0, 0, 1]
     )
-    root_part.create_body(name="Body.1")
+    body1 = root_part.create_body(name="Body.1")
+    body1.create_face(name="TheFaceF").set_vertices([0, 0, 0, 1, 0, 0, 0, 1, 0]).set_facets(
+        [0, 1, 2]
+    ).set_normals([0, 0, 1, 0, 0, 1, 0, 0, 1])
     assert root_part.part_link is None
     assert p.scene_link.get().part_guid == ""
 
@@ -341,7 +355,10 @@ def test_reset_part(speos: Speos):
     sp1.create_sub_part(name="SubPart.11").set_axis_system(
         axis_system=[-5, -4, -10, 1, 0, 0, 0, 1, 0, 0, 0, 1]
     )
-    root_part.create_body(name="Body.1")
+    body1 = root_part.create_body(name="Body.1")
+    body1.create_face(name="TheFaceF").set_vertices([0, 0, 0, 1, 0, 0, 0, 1, 0]).set_facets(
+        [0, 1, 2]
+    ).set_normals([0, 0, 1, 0, 0, 1, 0, 0, 1])
     root_part.commit()
     assert root_part.part_link is not None
 
@@ -368,7 +385,10 @@ def test_delete_part(speos: Speos):
     sp1.create_sub_part(name="SubPart.11").set_axis_system(
         axis_system=[-5, -4, -10, 1, 0, 0, 0, 1, 0, 0, 0, 1]
     )
-    root_part.create_body(name="Body.1")
+    body1 = root_part.create_body(name="Body.1")
+    body1.create_face(name="TheFaceF").set_vertices([0, 0, 0, 1, 0, 0, 0, 1, 0]).set_facets(
+        [0, 1, 2]
+    ).set_normals([0, 0, 1, 0, 0, 1, 0, 0, 1])
     root_part.commit()
     assert root_part.part_link is not None